requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
requests==2.31.0
requests-mock==1.11.0
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
